Fix crash when new folder is created
Turns out the problem is with Core Data assigning a temporary ID to
new entities. The CollectionView was filled temporary IDs, causing the
app to crash when existingObject(with:) was called because the temporary
ID had changed to a permanent one. We solved this by forcing Core Data
to assign a permanent ID at an entity's initialization.