Add rename folder functionality
I found out that we do not to call reloadItems() when directly using
a snapshot given by a FetchedResultsController. We can just apply the
snapshot and it'll work. However, when using multiple groups like in
the Sidebar View, we have to call it. I'm guessing because we have to
tell the snapshot that is being applied to reload the items.